Move formatting strings code to separate file, ignore in coverage
