+++ codecov.yml
@@ -1,0 +1,11 @@
+ignore:
+  - napari_hub_cli/formatting.py
+coverage:
+  status:
+    project:
+      default:
+        target: auto
+        threshold: 0.5%  # coverage can drop by up to 0.5% while still posting success
+    patch: off
+comment:
+  require_changes: true  # if true: only post the PR comment if coverage changes
